+++ src/playground/views/main/index.js
@@ -46,20 +46,14 @@ const styles = theme => ({
   },
 });
 
-const infoSettings = {
-  info : {
-    timeout: null,
-    showCloseButton: true,
-    showActionButton: true,
-    showProgressBar: true,
-    enableClickAwayListener: true,
-    actionText: 'Click Me!!!',
-    actionClickListener: () => { console.log('Listener triggered!'); },
-  },
-};
-
-const longText = 'This is an unnecessarily long message to the user!!!!';
-
+/**
+ * NOTE ABOUT DEVELOPMENT PLAYGROUND:
+ * The playground is setup to instantiate alerts directly through the AlertsConsumer (as seen below).
+ * This is considered advanced usage in the documentation, and does give you more control over the
+ * alerts. There is also an Alert component setup for you at the top of the render function. To use
+ * the Alert component, just uncomment the onClick on line 102, and comment out the onClick on lines
+ * 99-101 below. This is a good way to get an understanding of how the two approaches work.
+ */
 class Main extends Component {
   state = {
     open: false,
@@ -82,7 +76,7 @@ class Main extends Component {
           showProgressBar
           enableClickAwayListener
           actionText={'text'}
-          actionClickListener={() => {}}
+          actionClickListener={() => console.log('Clicked!')}
         />
         <AlertsConsumer>
           {context => {
@@ -91,29 +85,41 @@ class Main extends Component {
                 <Button
                   className={classes.info}
                   variant="contained"
-                  onClick={() => context.info('Header', longText, null, infoSettings)}
+                  onClick={() =>
+                    context.info(
+                      'Info Header',
+                      'Info message! -- default theme and settings')}
                 >
                   Info
                 </Button>
                 <Button
                   className={classes.success}
                   variant="contained"
-                  // onClick={() => context.success('Header', 'message!')}
-                  onClick={() => this.setState({ open: true })}
+                  onClick={() => 
+                    context.success(
+                    'Yellow Success Header',
+                    'White success message! -- no adornment, green body background, 7 second timeout')}
+                  // onClick={() => this.setState({ open: true })}
                 >
                   Success
                 </Button>
                 <Button
                   className={classes.warning}
                   variant="contained"
-                  onClick={() => context.warning(null, 'message!')}
+                  onClick={() => 
+                    context.warning(
+                      null,
+                      'White warning message! -- null header, orange body background, click away listener enabled, 10 second timeout')}
                 >
                   Warning
                 </Button>
                 <Button
                   className={classes.error}
                   variant="contained"
-                  onClick={() => context.error('Header', null)}
+                  onClick={() =>
+                    context.error(
+                      'Error Header',
+                      'Error message! -- progress bar enabled, 10 second timeout, close button enabled')}
                 >
                   Error
                 </Button>