+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 0.7.2
+- [fix] Add padding to body when no action or close button present.
+- [fix] Add border radius to body when no adornment is present.
+- [fix] Body color specificity.
+
 ## 0.7.1
 - [fix] Update TypeScript definitions for new wrapper.
 - [fix] Update README for new wrapper.